Fluids - fix unused function warnings
