+++ examples/fluids/qfunctions/setupboundary.h
@@ -15,7 +15,7 @@
 // testbed platforms, in support of the nation's exascale computing imperative.
 
 /// @file
-/// Geometric factors for boundary integral in Navier-Stokes example using PETSc
+/// Geometric factors (3D) for boundary integral in Navier-Stokes example using PETSc
 
 #ifndef setupboundary_h
 #define setupboundary_h
@@ -56,7 +56,7 @@
 // *****************************************************************************
 CEED_QFUNCTION(SetupBoundary)(void *ctx, CeedInt Q,
                               const CeedScalar *const *in, CeedScalar *const *out) {
-  // *INDENT-OFF*                              
+  // *INDENT-OFF*
   // Inputs
   const CeedScalar (*J)[3][CEED_Q_VLA] = (const CeedScalar(*)[3][CEED_Q_VLA])in[0],
                    (*w) = in[1];
@@ -89,62 +89,6 @@ CEED_QFUNCTION(SetupBoundary)(void *ctx, CeedInt Q,
     q_data_sur[2][i] = J2 / detJb;
     q_data_sur[3][i] = J3 / detJb;
 
-  } // End of Quadrature Point Loop
-
-  // Return
-  return 0;
-}
-
-// *****************************************************************************
-// This QFunction sets up the geometric factor required for integration when
-//   reference coordinates are in 1D and the physical coordinates are in 2D
-//
-// Reference (parent) 1D coordinates: X
-// Physical (current) 2D coordinates: x
-// Change of coordinate vector:
-//           J1 = dx_1/dX
-//           J2 = dx_2/dX
-//
-// detJb is the magnitude of (J1,J2)
-//
-// All quadrature data is stored in 3 field vector of quadrature data.
-//
-// We require the determinant of the Jacobian to properly compute integrals of
-//   the form: int( u v )
-//
-// Stored: w detJb
-//   in q_data_sur[0]
-//
-// Normal vector is given by the cross product of (J1,J2)/detJ and ẑ
-//
-// Stored: (J1,J2,0) x (0,0,1) / detJb
-//   in q_data_sur[1:2] as
-//   (detJb^-1) * [ J2 ]
-//                [-J1 ]
-//
-// *****************************************************************************
-CEED_QFUNCTION(SetupBoundary2d)(void *ctx, CeedInt Q,
-                                const CeedScalar *const *in, CeedScalar *const *out) {
-  // *INDENT-OFF*
-  // Inputs
-  const CeedScalar (*J)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[0],
-                   (*w) = in[1];
-  // Outputs
-  CeedScalar (*q_data_sur)[CEED_Q_VLA] = (CeedScalar(*)[CEED_Q_VLA])out[0];
-  // *INDENT-ON*
-
-  CeedPragmaSIMD
-  // Quadrature Point Loop
-  for (CeedInt i=0; i<Q; i++) {
-    // Setup
-    const CeedScalar J1 = J[0][i];
-    const CeedScalar J2 = J[1][i];
-
-    const CeedScalar detJb = sqrt(J1*J1 + J2*J2);
-
-    q_data_sur[0][i] = w[i] * detJb;
-    q_data_sur[1][i] = J2 / detJb;
-    q_data_sur[2][i] = -J1 / detJb;
   } // End of Quadrature Point Loop
 
   // Return
